Change in BC family name, exclusion of CH, change in girder definition.
The dipole BC is divided in two families name in the AT model. One is called
bc_lf (bc low field) and the other bc_hf (bc high field).
We excluded the horizontal corrector in the sd2j and sd2k sextupole families;
The way girders are defined in the lattice were changed; now there are elements which
do not belong to any girder, such as some drifts and the elements in the straight sections.
Also, the markers 'girder' are excluded from the model after the girders are assigned to the
elements.